refactor: extract shared utilities and split check module
- Extract validateMonth, currentMonth, parseAmount into month.ts
- Split check.ts (264 lines, 17 exports) into check.compute.ts and check.render.ts
- check.ts is now a slim orchestrator (~95 lines)
- Remove duplicate formatEuro/formatEuroDecimal tests from check.test.ts
- Move shared utility tests to month.test.ts
- Update all imports across budget command group
